requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Web Interface
streamlit==1.28.1

# RAG and AI Framework
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# OpenAI API
openai>=1.6.1,<2.0.0

# Vector Database
faiss-cpu==1.7.4

# PDF Processing
PyPDF2==3.0.1

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Requests (for Streamlit to FastAPI communication)
requests==2.31.0

# File Handling
python-multipart==0.0.6

# Optional: For better performance
numpy==1.24.3
pandas==2.0.3
langchain-openai
